Need to use _wstat instead of _stat to stat files on windows
_stat doesn't handle multibyte character sequences, unless (afaict)
the current codepage supports them, and then it uses the current
codepage.  In order to correctly handle multibyte characters in a
file path we have to use _wstat.

In deference to the fact that the CF implementation may be used on
platforms other than windows i've added fileSize to the FileSystem
helper functions, and added a windows impl.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@25577 268f45cc-cd09-0410-ab3c-d52691b4dbfc