Rubberstamped by Eric.
Move current layout tests results from platform/mac/svg to platform/mac-tiger/svg.
The problem is that the current results have accidently been generated with Leopard,
I'm still moving this to mac-tiger so we don't loose the history for all the results.
After this commit, the baseline will be regenerated in mac-tiger/svg, and someone with
Leopard needs to generate new results in mac-leopard/svg.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@26597 268f45cc-cd09-0410-ab3c-d52691b4dbfc