Replace post-parse pass to find declarations with logic in the parser itself
Reviewed by Geoff.

Instead of finding declarations in a pass following the initial parsing of
a program, we incorporate the logic directly into the parser.  This lays
the groundwork for further optimisations (such as improving performance in
declaration expressions -- var x = y; -- to match that of standard assignment)
in addition to providing a 0.4% performance improvement in SunSpider.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@28854 268f45cc-cd09-0410-ab3c-d52691b4dbfc