Lars Knoll <lars@trolltech.com>
Fix rendering of the Scrollbar as well as mouse handling for some styles.

The QStyle expects that that painter is set up to clip to the scrollbar bounds
and some of the styles seem to paint somewhat outside of these bounds. Clipping to
the scrollbar bounds before drawing removes some artifacts.

Also set m_opt.rect.topLeft to (0/0) in the mouse handlers. Fixes some issues with
the oxygen style.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@29736 268f45cc-cd09-0410-ab3c-d52691b4dbfc