Make canvas use an ImageBuffer for its backing store
Reviewed by Alp Toker.

In order to make the canvas implementation less platform dependent
(and thus reduce the current quagmire of ifdefs) we now use an
ImageBuffer to provide the backing buffer, an immediate consequence
of this is to remove multiple ifdefs in the construction of the
buffer.  This patch allows us to further reduce the platform
dependencies in later patches.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@30163 268f45cc-cd09-0410-ab3c-d52691b4dbfc