Bug 17486: Support HTML5 Canvas.getImageData API Support Canvas.getImageData and Canvas.createImageData
Reviewed by Sam W.

This patch adds support for all the pixel reading portions
of the HTML5 Canvas spec.  There are two new types ImageData
and CanvasPixelArray which are used to provide the HTML5
ImageData object, and the required semantics for assignment
to the ImageData data array.

We only implement the CG version of ImageBuffer::getImageData,
but the logic is null safe, so this will not introduce any
crashes into other platforms, unfortunately it will result in
JS Object detection "lying" on non-CG platforms.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@30506 268f45cc-cd09-0410-ab3c-d52691b4dbfc