r25512 inlined the urlIcon to improve the startup time of Safari on Mac. Inlining is not suitable for platforms where this icon is themable. E.g. the Qt platform is allowing to theme this icon and the Gtk+ platform will probably end up with themable icons as well.
Remove dead code from the windows port and move the urlIcon from the windows
directory to the Resource directory to be used by Qt and other ports.

Ifdef the usage of the built-in icon in IconDatabase.cpp and for Qt do not use
it. Gtk+ currently has no proper implementation of Image::loadPlatformResource but
once it does it wants to use this for getting the urlIcon as well.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@30961 268f45cc-cd09-0410-ab3c-d52691b4dbfc