Holger Hans Peter Freyther  <zecke@selfish.org>
Fix crash when setHtml gets called from slots connected to QWebPage::unsupportedContent

Ben found this crash in the demo browser.

In void MainResourceLoader::continueAfterContentPolicy(PolicyAction contentPolicy,

case PolicyDownload:
    frameLoader()->client()->download(m_handle.get(), request(), m_handle.get()->request(), r);
    receivedError(interruptionForPolicyChangeError());

We are in a policy check, the download() call will call setHtml, which will start loading
the data into a frame. The loading will cancel all policy checks and call ResourceLoader::releaseResources
which will set m_frame to 0. Then we return and invoke interruptionForPolicyChangeError() which
will use the m_frame->loader() but it is gone. Do not call into receivedError and
interruptionForPolicyChangeError if we are gone anyway...


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@32688 268f45cc-cd09-0410-ab3c-d52691b4dbfc