Fix builds using the "debug" variant.
Rubber-stamped by Oliver Hunt.

This reverts r36130 and tweaks Identifier to export the same symbols for Debug
and Release configurations.

* Configurations/JavaScriptCore.xcconfig:
* DerivedSources.make:
* JavaScriptCore.Debug.exp: Removed.
* JavaScriptCore.base.exp: Removed.
* JavaScriptCore.exp: Added.
* JavaScriptCore.xcodeproj/project.pbxproj:
* kjs/identifier.cpp:
(JSC::Identifier::addSlowCase): #ifdef the call to checkSameIdentifierTable so that
there is no overhead in Release builds.
(JSC::Identifier::checkSameIdentifierTable): Add empty functions for Release builds.
* kjs/identifier.h:
(JSC::Identifier::add): #ifdef the calls to checkSameIdentifierTable so that there is
no overhead in Release builds, and remove the inline definitions of checkSameIdentifierTable.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@36291 268f45cc-cd09-0410-ab3c-d52691b4dbfc