<https://bugs.webkit.org/show_bug.cgi?id=21414> REGRESSION: Regular Expressions and character classes, shorthands and ranges <rdar://problem/6543487>
Reviewed by Gavin Barraclough.

In certain circumstances when WREC::Generator::generateCharacterClassInvertedRange invokes
itself recursively, it will incorrectly emit (and thus consume) the next single character
match in the current character class.  As WREC uses a binary search this out of sequence
codegen could result in a character match being missed and so cause the regex to produce
incorrect results.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@40522 268f45cc-cd09-0410-ab3c-d52691b4dbfc