Fix <https://bugs.webkit.org/show_bug.cgi?id=23858> Bug 23858: Crash when removing a HTMLSelectElement from the document from inside its focus event handler
Reviewed by Darin Adler.

* html/HTMLSelectElement.cpp:
(WebCore::HTMLSelectElement::menuListDefaultEventHandler): Don't store the renderer in a local variable
as it can be invalidated by any of the calls to focus() within the function.  Instead, retrieve it and
null-check it when it is needed.

Test for <https://bugs.webkit.org/show_bug.cgi?id=23858>
Bug 23858: Crash when removing a HTMLSelectElement from the document from inside its focus event handler

Reviewed by Sam Weinig.

* fast/dom/HTMLSelectElement/remove-element-from-within-focus-handler-crash-expected.txt: Added.
* fast/dom/HTMLSelectElement/remove-element-from-within-focus-handler-crash.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@40804 268f45cc-cd09-0410-ab3c-d52691b4dbfc