<rdar://problem/6156755> onMouseOver events do not fire properly for cross frame drag and drop
Reviewed by Adele Peterson.

This problem was caused by incorrectly ignoring whether or not the
default behaviour of the mousedown event was suppressed.  If a
mousedown handler in a frame prevents default handling then the
subsequent mousemove events fired for the drag should not be
captured by that frame, should the mouse move out of its bounds.

Test: fast/events/mouse-drag-from-frame.html


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@40845 268f45cc-cd09-0410-ab3c-d52691b4dbfc