[Gtk] add options for 3D transforms and HTML5 channel messaging to the build https://bugs.webkit.org/show_bug.cgi?id=24072
Reviewed by Alexey Proskuryakov.

Add options for toggling 3D transforms and HTML5 channel messaging
support on or off.

Also fix the web-workers option. It should be web-workers and not workers.

    * configure.ac:

Allow toggling of 3D transforms and HTML5 channel messaging
support for the Autotools (Gtk) build.

Also add '--gtk' in the build-webkit help doc and fix autotools
option for web-workers support.

    * Scripts/build-webkit:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@41172 268f45cc-cd09-0410-ab3c-d52691b4dbfc