<rdar://problem/6681868> When building with Xcode 3.1.3 should be using gcc 4.2
Rubber-stamped by Darin Adler.

The meaning of XCODE_VERSION_ACTUAL is more sensible in newer versions of Xcode.
Update our logic to select the compiler version to use the more appropriate XCODE_VERSION_MINOR
if the version of Xcode supports it, and fall back to XCODE_VERSION_ACTUAL if not.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43691 268f45cc-cd09-0410-ab3c-d52691b4dbfc