Fix <https://bugs.webkit.org/show_bug.cgi?id=25785>. Bug 25785: Segfault in mark when using JSObjectMakeConstructor
Reviewed by Oliver Hunt.

* API/JSObjectRef.cpp:
(JSObjectMakeConstructor): OpaqueJSClass::prototype can return 0.  We need to use the default object prototype when it does.
* API/tests/testapi.c:
(main): Add a test case.
* runtime/JSObject.h:
(JSC::JSObject::putDirect): Add a clearer assertion for a null value.  The assertion on the next line does catch this,
but the cause of the failure is not clear from the assertion itself.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43692 268f45cc-cd09-0410-ab3c-d52691b4dbfc