[GTK] Implement Image::loadPlatformResource.
Install the four resources we are using. For the icons
where an icon name is specified by freedeskop.org try to
map the WebCore name to that name and fallback to the
internal icon. For everything else use the internal icon
name. Remove the temporary link stub as this is not
required any more.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43955 268f45cc-cd09-0410-ab3c-d52691b4dbfc