Fix <https://bugs.webkit.org/show_bug.cgi?id=25917> / <rdar://problem/6910066>. Bug 25917: REGRESSION (r43559?): Javascript debugger crashes when pausing page
Reviewed by Oliver Hunt.

The debugger currently retrieves the arguments object from an activation rather than pulling
it from a call frame.  This is unreliable to due to the recent optimization to lazily create
the arguments object.  In the long-term it should stop doing that (<rdar://problem/6911886>),
but for now we force eager creation of the arguments object when debugging.

* bytecompiler/BytecodeGenerator.cpp:
(JSC::BytecodeGenerator::BytecodeGenerator):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@43976 268f45cc-cd09-0410-ab3c-d52691b4dbfc