Bug 26249: Support JSON.stringify <https://bugs.webkit.org/show_bug.cgi?id=26249>
Reviewed by Sam Weinig.

Implement JSON.stringify.  This patch handles all the semantics of the ES5
JSON.stringify function, including replacer functions and arrays and both
string and numeric gap arguments.

Currently uses a clamped recursive algorithm basically identical to the spec
description but with a few minor tweaks for performance and corrected semantics
discussed in the es-discuss mailing list.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@44521 268f45cc-cd09-0410-ab3c-d52691b4dbfc