Make LiteralParser non-recursive
Reviewed by Geoff Garen.

Convert LiteralParser from using a simple recursive descent parser
to a hand rolled PDA.  Relatively simple conversion, but required
modifications to MarkedArgumentBuffer to make it more suitable as
a generic marked vector.   I'll refactor and rename MarkedArgumentBuffer
in future as there are many other cases where it will be useful to
have such a class.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@44644 268f45cc-cd09-0410-ab3c-d52691b4dbfc