Fix <http://webkit.org/b/28484> Plug-in-related leaks seen on the build bot
Reviewed by Dan Bernstein.

WebKit/mac:

When the plug-in data is being loaded manually there is a reference cycle between the
NetscapePluginInstanceProxy and the HostedNetscapePluginStream.  We need to explicitly
break the reference cycle in NetscapePluginInstanceProxy::cleanup so that both objects
will be destroyed.

Take the opportunity to add RefCountedLeakCounter support to HostedNetscapePluginStream
and NetscapePluginInstanceProxy to simplify tracking down leaks of these objects in the future.

* Plugins/Hosted/HostedNetscapePluginStream.h:
* Plugins/Hosted/HostedNetscapePluginStream.mm:
(WebKit::HostedNetscapePluginStream::HostedNetscapePluginStream):
(WebKit::HostedNetscapePluginStream::~HostedNetscapePluginStream):
* Plugins/Hosted/NetscapePluginInstanceProxy.mm:
(WebKit::NetscapePluginInstanceProxy::NetscapePluginInstanceProxy):
(WebKit::NetscapePluginInstanceProxy::~NetscapePluginInstanceProxy):
(WebKit::NetscapePluginInstanceProxy::cleanup): Clear m_manualStream to break the reference
cycle.

WebKitTools:

Update check-for-global-initializers to accommodate the new uses of RefCountedLeakCounter in WebKit.

* Scripts/check-for-global-initializers:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@47551 268f45cc-cd09-0410-ab3c-d52691b4dbfc