[Qt/Mac] Un-blacklist the QuickTime plugin
Patch by Cédric Luthi <cedric.luthi@gmail.com> on 2009-08-28
Reviewed by Tor Arne Vestbø

https://bugs.webkit.org/show_bug.cgi?id=28768

Now that the drawing model is properly negotiated, there is no need the
blacklist the QuickTime plugin anymore.

This revealed a bug in PluginView::performRequest: it should not initiate a
request if the plugin has been stopped, which is what happens when a plugin
only supports the QuickDraw drawing model.

* plugins/PluginView.cpp:
* plugins/mac/PluginPackageMac.cpp:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@47863 268f45cc-cd09-0410-ab3c-d52691b4dbfc