WebCore: Removed some vestigial and potentially crashy code that makes consolidating event listeners hard.
Patch by Geoffrey Garen <ggaren@apple.com> on 2009-09-16
Reviewed by Sam Weinig and Anders Carlsson.

* dom/Node.cpp:
(WebCore::Node::addEventListener): No need to limit when event listeners
can be added. Other EventTargets don't do this.
        
Also, check for null when accessing document(). Technically, the JS bindings
do this check for us, but let's not rely on that.

LayoutTests: Added a test for a potential crash when adding an event listener to a
doctype node.

Patch by Geoffrey Garen <ggaren@apple.com> on 2009-09-16
Reviewed by Sam Weinig.

* fast/dom/doctype-event-listener-crash-expected.txt: Added.
* fast/dom/doctype-event-listener-crash.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@48447 268f45cc-cd09-0410-ab3c-d52691b4dbfc