Allow [Reflect] on SVG elements. https://bugs.webkit.org/show_bug.cgi?id=28936
Reviewed by Sam Weinig.

Update the JS binding generators to reference SVGNames instead of
HTMLNames, if [Reflect]ing an attribute on an SVG element.  Make
SVGElement::id use [Reflect].

Also make [Reflect] on an attribute with a setter exception work in ObjC
bindings.

WebCore: 

Test: svg/dom/id-reflect.html

* bindings/scripts/CodeGenerator.pm: Add a function to determine the
appropriate C++ namespace for attribute name constants.
* bindings/scripts/CodeGeneratorObjC.pm: Generate ExceptionCode handling
code for [Reflect] on an attribute with a setter exception.
* bindings/scripts/CodeGeneratorCOM.pm: Generate "SVGNames" instead of
"HTMLNames" when appropriate.
* bindings/scripts/CodeGeneratorJS.pm: Ditto.
* bindings/scripts/CodeGeneratorV8.pm: Ditto.
* svg/SVGElement.cpp: Remove getter and setter methods for id.
* svg/SVGElement.h: Ditto.
* svg/SVGElement.idl: Add [Reflect] to id.

LayoutTests: 

* svg/dom/script-tests/id-reflect.js: Added.
* svg/dom/id-reflect-expected.txt: Added.
* svg/dom/id-reflect.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@49424 268f45cc-cd09-0410-ab3c-d52691b4dbfc