Unreviewed documentation fix from David Boddie (Qt Doc Team)
Patch by Kenneth Rohde Christiansen <kenneth@webkit.org> on 2009-10-26
Ensure that qdoc will always see the RenderHints property.

The property was only defined in the header file if the Q_OS_SYMBIAN
symbol was not defined, resulting in the property not showing up
in the Qt documentation just because one platform doesn't support it.

A follow up commit will improve the documentation for the property
and note that it is not supported on the Symbiam platform.

* Api/qwebview.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@50096 268f45cc-cd09-0410-ab3c-d52691b4dbfc