Fix assertion in SVGRenderBase::mapLocalToContainer resulting from HitTestResult::boundinBox()
https://bugs.webkit.org/show_bug.cgi?id=27347

The current HitTestResult methods are not using the (3d)
transformation aware routines. This can lead to an assertion
SVGRenderBase::mapLocalToContainer method.

Remove HitTestResult::boundingBox() as it is only used in
two places and conceptually doesn't belong into a HitTest
which is operating on points. Update the Qt and Win code
to determine the rect themselves in a (3d) transformation
fashion.

Change HitTestResult::imageRect to use the (3d) transformation
aware RenderBox::absoluteContentQuad to avoid running into
an assertion with SVG content.

* rendering/HitTestResult.cpp:
(WebCore::HitTestResult::imageRect):
* rendering/HitTestResult.h:

* WebElementPropertyBag.cpp:
(WebElementPropertyBag::Read):

* Api/qwebframe.cpp:
(QWebHitTestResultPrivate::QWebHitTestResultPrivate):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@50207 268f45cc-cd09-0410-ab3c-d52691b4dbfc