REGRESSION (r50218-r50262): E*TRADE accounts page is missing content https://bugs.webkit.org/show_bug.cgi?id=30947 <rdar://problem/7348833>
Reviewed by Maciej Stachowiak

The logic for flagging that a structure has non-enumerable properties
was in addPropertyWithoutTransition, rather than in the core Structure::put
method.  Despite this I was unable to produce a testcase that caused
the failure that etrade was experiencing, but the new assertion in
getEnumerablePropertyNames triggers on numerous layout tests without
the fix, so in effecti all for..in enumeration in any test ends up
doing the required consistency check.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@50320 268f45cc-cd09-0410-ab3c-d52691b4dbfc