Possible crash in RenderSlider::layout. https://bugs.webkit.org/show_bug.cgi?id=31016
Patch by David Levin <levin@chromium.org> on 2009-11-02
Reviewed by Dan Bernstein.

Fix out of place line of code.

Test: scrollbars/overflow-scrollbar-combinations.html
This crash only seems to repro when WebKit draws the play controls,
so the crash repros in chromium running this test but not WebKit
nightlies which use QuickTime to draw the controls.

* rendering/RenderSlider.cpp:
(WebCore::RenderSlider::layout):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@50417 268f45cc-cd09-0410-ab3c-d52691b4dbfc