DOMHTMLOptionsCollection is missing some implementation.
https://bugs.webkit.org/show_bug.cgi?id=31488

Reviewed by Dan Bernstein.

* DOMHTMLClasses.cpp:
(DOMHTMLOptionsCollection::DOMHTMLOptionsCollection):
Initialize m_collection.
(DOMHTMLOptionsCollection::createInstance):
Create a DOMHTMLOptionsCollection. If we fail to query for
IDOMHTMLOptionsCollection, delete it, and return 0. Otherwise, return
the result.
(DOMHTMLOptionsCollection::length):
(DOMHTMLOptionsCollection::item):
Create a DOMNode for the WebCore Node. If this is 0, return E_FAIL.
(DOMHTMLOptionsCollection::namedItem):
Correct the signature of this function.

* DOMHTMLClasses.h:
Declare DOMHTMLOptionsCollection::createInstance(). Correct the
signature of namedItem() to match IDOMHTMLOptionsCollection. Add a
member to DOMHTMLOptionsCollection to hold the WebCore object.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@51091 268f45cc-cd09-0410-ab3c-d52691b4dbfc