<rdar://problem/7458272> REGRESSION (r50470) Reproducible ASSERT in RenderWidget::setWidgetGeometry()
Reviewed by Darin Adler.

WebCore: 

Test: fast/forms/defer-updateFromElement.html

* html/HTMLFormControlElement.cpp:
(WebCore::updateFromElementCallback): Added this post-attach callback
that invokes updateFromElement().
(WebCore::HTMLFormControlElement::recalcStyle): Instead of calling
updateFromElement(), enqueue a post-attach (in this case, post-style
recalc) callback to do it. updateFromElement() can cause the selection
to change, and in turn trigger synchronous layout, so it must not be
called during style recalc.

LayoutTests: 

* fast/forms/defer-updateFromElement-expected.txt: Added.
* fast/forms/defer-updateFromElement.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@52100 268f45cc-cd09-0410-ab3c-d52691b4dbfc