Part of <http://webkit.org/b/32568>. Web Inspector: Context Menus should be used in more places.
Reviewed by Timothy Hatcher.

Add a context menu entry in the console to clear the console. This
is only shown when there is no user selection of text, because if
there is selection of the text, then we want to give the user the Copy +
Spell Check context menu that they are used to.

* English.lproj/localizedStrings.js: Added "Clear Console".
* inspector/front-end/ConsoleView.js:
(WebInspector.ConsoleView.prototype._handleContextMenuEvent):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@52217 268f45cc-cd09-0410-ab3c-d52691b4dbfc