WebCore: Right clicking in a table cell creates a selection that extends into the next cell. <rdar://problem/5565465> https://bugs.webkit.org/show_bug.cgi?id=32622
Reviewed by Eric Seidel.

Avoid extending the selection to the next cell if we start from an empty cell.

Test: editing/selection/empty-cell-right-click.html

* editing/TypingCommand.cpp:
(WebCore::TypingCommand::deleteKeyPressed): Replaced the code that check for empty table cell
with a call to the new function.
* editing/VisibleSelection.cpp:
(WebCore::VisibleSelection::setStartAndEndFromBaseAndExtentRespectingGranularity): Extends the
selection only if we are not starting from an empty table cell.
* editing/htmlediting.cpp:
(WebCore::isEmptyTableCell): Added a function since this check is performed now in two places.
* editing/htmlediting.h: Added isEmptyTableCell declaration.

LayoutTests: Right clicking in a table cell creates a selection that extends into the next cell.
<rdar://problem/5565465>
https://bugs.webkit.org/show_bug.cgi?id=32622

Reviewed by Eric Seidel.

* editing/selection/empty-cell-right-click-expected.txt: Added.
* editing/selection/empty-cell-right-click.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@52220 268f45cc-cd09-0410-ab3c-d52691b4dbfc