MSAA: Accessibility role of list items is wrong
https://bugs.webkit.org/show_bug.cgi?id=32688

Reviewed by Adam Roben.

WebCore:

Test: platform/win/accessibility/list-item-role.html

* accessibility/AccessibilityObject.h:
(WebCore::):
Add a ListItemRole to the enum.

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::msaaRoleForRenderer):
Return the role for the RenderObject, or UnknownRole if unknown.
(WebCore::AccessibilityRenderObject::roleValueForMSAA):
Call msaaRoleForRenderer() to get the role. If it is UnknownRole, fall
back to m_role.

WebKit/win:

* AccessibleBase.cpp:
(MSAARole):
Make the WebCore list item role map to the MSAA list item role.

LayoutTests:

* platform/win/accessibility/list-item-role-expected.txt: Added.
* platform/win/accessibility/list-item-role.html: Added.
Verify that the list item role is "list item".

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@52351 268f45cc-cd09-0410-ab3c-d52691b4dbfc