WebKit returns incorrect match count when first string match is on an element with user-select-none style https://bugs.webkit.org/show_bug.cgi?id=33508 <rdar://problem/7482143>
Reviewed by Darin Adler.

If we use VisiblePosition to calculate the start of the new search range,
we skip all the nodes with user-select-none style, producing an incorrect
match count.
        
No new tests: there are no API's available to Javascript to verify the
number of matches found.

* page/Frame.cpp:
(WebCore::Frame::markAllMatchesForText): Changed to use the end of the previous
result range as the start of the new search range.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@53142 268f45cc-cd09-0410-ab3c-d52691b4dbfc