[Qt/Win] Flash in QGraphicsWebView does not process hover correctly.
Reviewed by Simon Hausmann.

Mouse hover does not work as expected with the flash in some sites.
    - http://www.bbc.co.uk/ Hover over the map
    - http://www.barbie.com/ Hover over the menu items (Games, Videos)
The problem appears to be that Flash queries NPNVnetscapeWindow on every
mouse hover. I do not how flash uses this value but returning 0 when flash
is in windowless mode solves the problem (When using QGraphicsWebView we
inject wmode opaque, thereby putting the plugin in windowless mode).

https://bugs.webkit.org/show_bug.cgi?id=33591

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@53327 268f45cc-cd09-0410-ab3c-d52691b4dbfc