[Qt] Fix incorrect dependency to QtXmlPatterns in generated include/QtWebKit/QtWebKit header
Patch by Thiago Macieira <thiago.macieira@nokia.com> on 2010-01-21
Reviewed by Simon Hausmann.

The generated file includes QtXmlPatterns/QtXmlPatterns, which is neither used/required by
the public QtWebKit API nor will it be available if Qt is configured with -no-xmlpatterns.

* WebCore.pro: Trick syncqt to believe that xmlpatterns is not a dependency, so that it's not
included in the generated file. It'll still be used and linked to with this trick.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@53627 268f45cc-cd09-0410-ab3c-d52691b4dbfc