[Android] [Qt] Touch event page co-ordinates are incorrect when touch is received in an iframe. https://bugs.webkit.org/show_bug.cgi?id=34162
Reviewed by Simon Hausmann.

WebCore: 

The pageX/pageY co-ordinates of a touch event should be relative to the containing frame, not the main frame. This change implements that and also updates the existing touch-inside-iframe test to also examine the page co-ordinates.

* dom/Touch.cpp:
(WebCore::Touch::Touch): Remove dead code.
* dom/Touch.h: Remove dead code.
* page/EventHandler.cpp:
(WebCore::EventHandler::handleTouchEvent): Rename framePoint to pagePoint, as it seems a clearer name. Also adjust the pagePoint to be relative to the touch target element's containing frame rather than the main frame.
* platform/PlatformTouchEvent.h: Rename Android specific constructor parameter to better reflect it's contents.
* platform/PlatformTouchPoint.h: Ditto.
* platform/android/PlatformTouchEventAndroid.cpp: Ditto.
* platform/android/PlatformTouchPointAndroid.cpp: Ditto.

LayoutTests: 

Update the existing iframe touch event test to also examine the page co-ordinates of the touch that is received.

* fast/events/touch/resources/touch-inside-iframe2.html:
* fast/events/touch/touch-inside-iframe-expected.txt:
* fast/events/touch/touch-inside-iframe.html:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@53994 268f45cc-cd09-0410-ab3c-d52691b4dbfc