Huge plain text pastes are slow <rdar://problem/5195056> https://bugs.webkit.org/show_bug.cgi?id=34237
Reviewed by Maciej Stachowiak.

WebCore: 

No new tests. It is a performance improvement.

The performance of the plain text paste is now linear.
We treat as a special case, a fragment that has been
created from plain text. Finding VisiblePositions and
applying style is greatly simplified, given the nature of
the fragment.

* editing/ReplaceSelectionCommand.cpp:
(WebCore::ReplaceSelectionCommand::doApply):
* editing/markup.cpp:
(WebCore::isPlainTextMarkup):
* editing/markup.h:

WebKit/mac: 

Mail is ignoring the fragment created from plain text or HTML that is
passed to the delegate function, that creates a new one.
This fix avoids creating the fragment twice.

* WebView/WebHTMLView.mm:
(-[WebHTMLView _pasteWithPasteboard:allowPlainText:]):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54036 268f45cc-cd09-0410-ab3c-d52691b4dbfc