Fix the Mac build.
Disable ENABLE_INDEXED_DATABASE since it is "completely non-functional".

As the comment in FeatureDefines.xcconfig notes, the list of feature defines
needs to be kept in sync across the various files.  The default values also
need to be kept in sync between these files and build-webkit.

JavaScriptCore:

* Configurations/FeatureDefines.xcconfig:

WebCore:

* Configurations/FeatureDefines.xcconfig:

WebKit/mac:

* Configurations/FeatureDefines.xcconfig:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54097 268f45cc-cd09-0410-ab3c-d52691b4dbfc