<rdar://problem/7044385> Crash at NetscapePlugInStreamLoader::didReceiveResponse()
Reviewed by Anders Carlsson.

The crash was caused by having two streams with a streamID of 1.

* Plugins/Hosted/NetscapePluginInstanceProxy.mm:
(WebKit::NetscapePluginInstanceProxy::NetscapePluginInstanceProxy): Fixed a typo. Now
correctly increments m_currentURLRequestID to account for the manual stream.
(WebKit::NetscapePluginInstanceProxy::disconnectStream): If the stream is the manual stream,
null it out instead of trying to remove it from the map. Added an assertion.
(WebKit::NetscapePluginInstanceProxy::loadRequest): Added an assertion.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54178 268f45cc-cd09-0410-ab3c-d52691b4dbfc