Bug 35065: Delay between page loading and animated GIF playing <http://webkit.org/b/35065> / <rdar://problem/7109548>
Reviewed by Darin Adler.

BitmapImage::startAnimation was adding the current frame duration to the desired start time
of the frame for every time it was called.  If the function then bailed out due to not having
sufficient data to render the frame, this would lead to the desired start time of the frame
being pushed out multiple times.  On an animated GIF that took mulitple seconds to load this
could happen many times for a single frame, resulting in the start time of the second frame
of the animation being pushed out by as much as five seconds.

* platform/graphics/BitmapImage.cpp:
(WebCore::BitmapImage::startAnimation): Change the order of the code slightly so that the
desired start time is only updated after determining that we have sufficient data to handle
the next frame.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54919 268f45cc-cd09-0410-ab3c-d52691b4dbfc