[Qt] Support private browsing mode in plugins
Add support for NPNVprivateModeBool property in plugins.

See also: https://developer.mozilla.org/En/Supporting_private_browsing_in_plugins

The NPNVprivateModeBool property is supported as scriptable property privateBrowsingEnabled
in the test WebKit plugin. The Mac platform also supports a cachedPrivateBrowsingEnabled
property implemented in the test plugin. This allows the Layout test plugins/private-browsing-mode.html
to retrieve the previous value of NPNVprivateModeBool in the test plugin. Due to the platform-specific
overhead required to support this bespoke property it is not implemented as part of this patch, instead
a new test, plugins/private-browsing-mode-2.html, is added to ensure that setting and resetting privateBrowsingEnabled works as expected.

http://bugs.webkit.org/show_bug.cgi?id=33180

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@55358 268f45cc-cd09-0410-ab3c-d52691b4dbfc