Bug 36083 - REGRESSION (r55772-r55834): Crash in JavaScriptCore RegExp code on PowerPC
Reviewed by Oliver Hunt, Darin Adler.

The problem is a bug in our port of PCRE - that a read may take place from the first character in an
empty string.  For the time being, revert to using a valid pointer in the data segment rather than
an invalid non-null pointer into the zero-page for the empty string's data pointer.  A better fix for
this will be to remove PCRE.

* runtime/UStringImpl.cpp:
(JSC::UStringImpl::empty):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56092 268f45cc-cd09-0410-ab3c-d52691b4dbfc