<http://webkit.org/b/37326> IDL files are being copied in to the WebCore framework again
Reviewed by Oliver Hunt.

WebCore:

* WebCore.xcodeproj/project.pbxproj: Wire up the script.  Remove the three inappropriate
files from the Copy Bundle Resources build phase.

WebKitTools:

Add a script to detect the presence of inappropriate files in the frameworks.  At present
it only looks for .css, .idl and .in files in the top level of WebCore.framework/Resources,
as these are the only cases we've encountered recently.  It can be extended to check the
other frameworks or for other inappropriate files in the future.

* Scripts/check-for-inappropriate-files-in-framework: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@57326 268f45cc-cd09-0410-ab3c-d52691b4dbfc