REGRESSION(r54368): Text drag-and-drop from input/textarea doesn't work if the text is like a URL https://bugs.webkit.org/show_bug.cgi?id=38374
Patch by Kent Tamura <tkent@chromium.org> on 2010-05-04
Reviewed by David Levin (and unofficially by Enrica Casucci).

WebCore:

Since r54368, documentFragmentFromDragData() creates an <a> element
with no anchor text for URL-like text dragged from input/textarea
element. If such text is dropped to input/textarea element, the text
is not inserted.
To fix this problem, use the original text or the URL as an anchor
text.

Test: editing/pasteboard/drag-drop-url-text.html

* page/DragController.cpp:
(WebCore::documentFragmentFromDragData):

LayoutTests:

* editing/pasteboard/drag-drop-url-text-expected.txt: Added.
* editing/pasteboard/drag-drop-url-text.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@58780 268f45cc-cd09-0410-ab3c-d52691b4dbfc