Fix the world.
In r59162 a change was made to WebCore's FeatureDefines.xcconfig that enabled FILE_READER and FILE_WRITER.
The author and reviewer of that patch ignored the carefully-worded warning at the top of that file asking
that changes to the file be kept in sync across JavaScriptCore, WebCore and WebKit, as well as being kept
in sync with build-webkit.  This led to WebCore and WebKit having different views of Document's vtable
and results in crashes in Safari shortly after launch when virtual function calls resulted in the wrong
function in WebCore being called.

We fix this by bringing the FeatureDefines.xcconfig files in to sync.  Based on the ChangeLog message and
other changes in r59162 it appears that enabling FILE_WRITER was unintentional so that particular change
has been reverted.

JavaScriptCore:

* Configurations/FeatureDefines.xcconfig:

WebCore:

* Configurations/FeatureDefines.xcconfig:

WebKit/mac:

* Configurations/FeatureDefines.xcconfig:

WebKit2:

* Configurations/FeatureDefines.xcconfig:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@59174 268f45cc-cd09-0410-ab3c-d52691b4dbfc