Spatial Navigation: refactor scrollInDirection to work with scrollable content https://bugs.webkit.org/show_bug.cgi?id=39195
Reviewed by Simon Fraser.
Patch by Antonio Gomes <tonikitoo@webkit.org>

WebCore:

scrollInDirection now receives as parameter the node that the Spatial Navigation
found as the more appropriated to move focus to. If it is in a scrollable container
(e.g. <div> with clipped overflow content), it scrolls recursively starting from
the container, not the current focused node.

Test: fast/events/spatial-navigation/snav-only-clipped-overflow-content.html

* page/FocusController.cpp:
(WebCore::FocusController::advanceFocusDirectionally):
* page/SpatialNavigation.cpp:
(WebCore::scrollInDirection):
* page/SpatialNavigation.h:

LayoutTests:

* fast/events/spatial-navigation/snav-only-clipped-overflow-content-expected.txt: Added.
* fast/events/spatial-navigation/snav-only-clipped-overflow-content.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@61298 268f45cc-cd09-0410-ab3c-d52691b4dbfc