[Qt] use "win32-g++*" scope to match all MinGW makespecs
Patch by Mark Brand <mabrand@mabrand.nl> on 2010-06-17
Reviewed by Simon Hausmann.

The scope "win32-g++" comes from the name of the makespec. However, it
is frequently used to check for MinGW. This works fine as long as
win32-g++ is the only makespec for MinGW. Now we need the wildcard
to cover "win32-g++-cross" as well.

.:

* WebKit.pri:

JavaScriptCore:

* JavaScriptCore.pro:

WebCore:

* WebCore.pro:

WebKit/qt:

* Api/DerivedSources.pro:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@61319 268f45cc-cd09-0410-ab3c-d52691b4dbfc