InsertOrderedList does not switch the list type properly when it has an inner list. https://bugs.webkit.org/show_bug.cgi?id=43166
Reviewed by Darin Adler.

WebCore: 

The bug was caused by forcedCreateList was not set to true when the start and the end
of the selection lies in the same list. Added selectionHasListOfType to fix this problem.

WebKit used not to convert the outer lists even when the list is fully selected.
Corrected this behavior by converting the entire list at once when the list is fully selected.
To decide whether or not a list is fully selected, added currentSelection argument to doApplyForSingleParagraph.

Tests: editing/execCommand/switch-list-type-with-inner-list.html
       editing/execCommand/switch-list-type-with-orphaned-li.html

* editing/InsertListCommand.cpp:
(WebCore::InsertListCommand::mergeWithNeighboringLists): Extracted the code to merge lists.
(WebCore::InsertListCommand::selectionHasListOfType): attachment.cgi
(WebCore::InsertListCommand::doApply): Calls selectionHasListOfType.
(WebCore::InsertListCommand::doApplyForSingleParagraph): See above.
(WebCore::InsertListCommand::listifyParagraph): Calls mergeWithNeighboringLists.
* editing/InsertListCommand.h:
* editing/htmlediting.cpp:
(WebCore::canMergeLists): Ensures lists being merged are instances of HTMLElement.
(WebCore::isNodeVisiblyContainedWithin): Works properly when one end is inside the range.

LayoutTests: 

Added a test to convert an unordered list with a nested list to an ordered list.
Also added a test to convert multiple lists with an orphaned list to an ordered list.
The behavior of WebKit is changed not to flatten the lists when inserting ordered/unordered
lists on nested lists.

* editing/execCommand/5207369-expected.txt: No longer flattens the list to match Firefox and MSIE.
* editing/execCommand/5207369.html: Updated.
* editing/execCommand/remove-list-items-expected.txt: Removed one whitespace.
* editing/execCommand/switch-list-type-with-inner-list-expected.txt: Added.
* editing/execCommand/switch-list-type-with-inner-list.html: Added.
* editing/execCommand/switch-list-type-with-orphaned-li-expected.txt: Added.
* editing/execCommand/switch-list-type-with-orphaned-li.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@64337 268f45cc-cd09-0410-ab3c-d52691b4dbfc