Handle ERROR_IO_INCOMPLETE more correctly in Connection
We still don't know exactly why we're getting this error, but at least
we can do something sensible when we do.

Fixes <http://webkit.org/b/44776> Occasional crash in
Connection::readEventHandler or assertion failure in
Connection::writeEventHandler due to ERROR_IO_INCOMPLETE

Reviewed by Sam Weinig.

* Platform/CoreIPC/win/ConnectionWin.cpp:
(CoreIPC::Connection::readEventHandler): Bail out of this function
entirely when we get ERROR_IO_INCOMPLETE. We'll get called back later
when the read completes. Continuing in the function at this point
would cause us to treat an incomplete read as a complete one, leading
to a crash. Added an assertion to make the crash more understandable
in the future.
(CoreIPC::Connection::writeEventHandler): Bail out when we get
ERROR_IO_INCOMPLETE. We'll get called back later when the write
completes.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@66231 268f45cc-cd09-0410-ab3c-d52691b4dbfc