<rdar://problem/7729077> Extending the selection to sentence boundary after a line break may select extra character https://bugs.webkit.org/show_bug.cgi?id=46232
Reviewed by Darin Adler.

WebCore: 

Test: editing/selection/extend-by-sentence-002.html

* editing/visible_units.cpp:
(WebCore::nextBoundary): The text iterator’s range end can be the position after
the line break, in which case the next visible is actually after the first character
of the next sentence. Instead, advance the text iterator past the newline character
and return the beginning of its range, which is guaranteed to still be before the
next sentence.

LayoutTests: 

* editing/selection/extend-by-sentence-002-expected.txt: Added.
* editing/selection/extend-by-sentence-002.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@67994 268f45cc-cd09-0410-ab3c-d52691b4dbfc