Crash when running plug-in layout tests https://bugs.webkit.org/show_bug.cgi?id=46269
Reviewed by Adam Roben.

Store the web page directly in the PluginView class, instead of getting it from the frame
of the plug-in element; this fixes two crashes, one where the plug-in element had been deallocated
before the plug-in view, and another where the plug-in element's document did not have a frame.

This also makes things more robust since we're asserting in the WebPage destructor that no plug-ins have
an outstanding reference to it.

* WebProcess/Plugins/PluginView.cpp:
(WebKit::webPage):
(WebKit::PluginView::PluginView):
(WebKit::PluginView::~PluginView):
(WebKit::PluginView::initializePlugin):
* WebProcess/Plugins/PluginView.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@68052 268f45cc-cd09-0410-ab3c-d52691b4dbfc