<rdar://problem/8478182> Make canHyphenate() return false for unsupported locales
Reviewed by John Sullivan.

No change in functionality, hence no new test.

* platform/text/cf/HyphenationCF.cpp:
(WebCore::::createValueForKey): Return 0 if hyphenation is not available for the locale.
(WebCore::cfLocaleCache): Moved the definition of the cache from lastHyphenLocation() to
this new getter to allow canHyphenate() to share it.
(WebCore::canHyphenate): Now returns true only if hyphenation is available.
(WebCore::lastHyphenLocation): Added an assertion.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@68334 268f45cc-cd09-0410-ab3c-d52691b4dbfc