document.nodesFromRect() needs to be removed from Document.idl https://bugs.webkit.org/show_bug.cgi?id=46492
Reviewed by Kenneth Christiansen and Andreas Kling.
Patch by Antonio Gomes <agomes@rim.com>

WebCore:

As discussed with Sam Weinig on IRC, it was decided to removed
nodesFromRect() from the Document.idl, and that way, not exposing
it to Web content.

Reason: there is no plan to make it standard at the moment, and at the
very least, the method should had been originally added with the "webkit"
prefix, which was not the case.

* dom/Document.idl:

LayoutTests:

Skip the tests relying on document.nodesFromRect in all platform for
the moment, and adjusting the expected results tests also accordingly.

* fast/dom/Window/window-properties-expected.txt:
* platform/gtk/Skipped:
* platform/mac/Skipped:
* platform/qt/Skipped:
* platform/qt/fast/dom/Window/window-properties-expected.txt:
* platform/win/Skipped:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@68345 268f45cc-cd09-0410-ab3c-d52691b4dbfc