Crash at lineBreakExistsAtPosition + 125 <rdar://problem/7028809> https://bugs.webkit.org/show_bug.cgi?id=46770
Reviewed by Darin Adler.

lineBreakExistsAtPosition is called from InsertTextCommand::input,
where we compute the downstream position of the endingSelection().
Downstream can return the original position, that comes from a VisibleSelection
but there is no guarantee that its renderer is still there. Everywhere we dereference
a renderer we check if it's null.
        
There is no regression test.
        
* editing/htmlediting.cpp:
(WebCore::lineBreakExistsAtPosition): Added check that the renderer is not null
before dereferencing it.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@68727 268f45cc-cd09-0410-ab3c-d52691b4dbfc