Crash at WebCore::nextCandidate + 27 https://bugs.webkit.org/show_bug.cgi?id=47118 <rdar://problem/7282934>
Reviewed by Darin Adler.

WebCore: 

When we canonicalize a Position to create a VisiblePosition, the position
is passed by reference. In canonicalPosition we call updateLayoutIgnorePendingStylesheets
that can produce a lot of side effects, including changing the selection.
This becomes a serious problem whne the position passed as reference is one of
the selection endpoints.
        
Test: editing/selection/focus-crash.html

* editing/VisiblePosition.cpp:
(WebCore::VisiblePosition::canonicalPosition):

LayoutTests: 

* editing/resources/iframebody.html: Added.
* editing/selection/focus-crash-expected.txt: Added.
* editing/selection/focus-crash.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@69051 268f45cc-cd09-0410-ab3c-d52691b4dbfc