Moving all bounding box related calculation to RenderSVGResourceFilterPrimitive https://bugs.webkit.org/show_bug.cgi?id=47174
Reviewed by Dirk Schulze.

The primitive sub-region calculation was spread in many
files across in WebKit source code, and all of them are moved
to RenderSVGResourceFilterPrimitive, where they should be.
In this patch the calculation is a static method which will
be turned to a real member function in the near future.

* platform/graphics/filters/FETile.cpp:
* platform/graphics/filters/FETile.h:
(WebCore::FETile::filterEffectType):
* platform/graphics/filters/Filter.h:
* platform/graphics/filters/FilterEffect.cpp:
* platform/graphics/filters/FilterEffect.h:
* rendering/RenderSVGResourceFilter.cpp:
(WebCore::RenderSVGResourceFilter::applyResource):
* rendering/RenderSVGResourceFilterPrimitive.cpp:
(WebCore::RenderSVGResourceFilterPrimitive::determineFilterPrimitiveSubregion):
* rendering/RenderSVGResourceFilterPrimitive.h:
(WebCore::RenderSVGResourceFilterPrimitive::RenderSVGResourceFilterPrimitive):
* svg/graphics/filters/SVGFilter.cpp:
* svg/graphics/filters/SVGFilter.h:
(WebCore::SVGFilter::mapLocalRectToAbsoluteRect):
(WebCore::SVGFilter::targetBoundingBox):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@70296 268f45cc-cd09-0410-ab3c-d52691b4dbfc