WebCore: Fix for https://bugs.webkit.org/show_bug.cgi?id=48814 <rdar://problem/8546143> Attempting to redo typing in apple.com/startpage search field causes infinite recursion in TextControlInnerTextElement::defaultEventHandler
Reviewed by Kent Tamura.

Test: editing/undo/redo-after-detach.html

* rendering/TextControlInnerElements.cpp: (WebCore::TextControlInnerTextElement::defaultEventHandler):
A TextControlInnerTextElement will almost always have a shadowAncestorNode, the HTMLInputElement.  In this case, 
after the renderer was destroyed, this shadow node was kept alive by the EditCommand, even though its not hooked up 
anymore to the shadow DOM.  EditCommands can sometimes operate on stale selections and are expected to fail silently.  
So here we prevent the infinite loop during event dispatch, and the rest of the redo operation will fail silently.

LayoutTests: Test for https://bugs.webkit.org/show_bug.cgi?id=48814
<rdar://problem/8546143> Attempting to redo typing in apple.com/startpage search field causes infinite recursion in TextControlInnerTextElement::defaultEventHandler

Reviewed by Kent Tamura.

* editing/undo/redo-after-detach-expected.txt: Added.
* editing/undo/redo-after-detach.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71131 268f45cc-cd09-0410-ab3c-d52691b4dbfc