WebProcessManager informed about crashes too late (pages may have already tried to load a new URL) https://bugs.webkit.org/show_bug.cgi?id=48853
Reviewed by Anders Carlsson.

This also fixes: <rdar://problem/8610451> Webkit2: Crash in WebKit!WebKit::WebProcessProxy::sendMessage opening google.com page

* UIProcess/WebProcessProxy.cpp:
(WebKit::WebProcessProxy::~WebProcessProxy):
Removed invalid assertion that is hit when the WebProcess dies unexpectedly.
(WebKit::WebProcessProxy::didClose):
Reordered so that the individual pages are informed that the processDidCrash() after the shared
WebProcessManager has been. This allows the pages to properly trigger a new process to be launched.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71135 268f45cc-cd09-0410-ab3c-d52691b4dbfc