Assertion failure in PluginStream::~PluginStream when running userscripts/user-script-plugin-document.html https://bugs.webkit.org/show_bug.cgi?id=48751 <rdar://problem/8615536>
Reviewed by Anders Carlsson.

WebKit/win: 

Always call committedLoad in WebFrameLoaderClient::finishedLoading, even if we have a manual loader. We were
running into a case where we were trying to load an empty plugin document, which uses a manual loader, and
PluginStream::didFinishLoading was never being called. The stream was never being stopped, making us fire
an assert in the PluginStream destructor.

* WebCoreSupport/WebFrameLoaderClient.cpp:
(WebFrameLoaderClient::finishedLoading):

LayoutTests: 

Unskip userscripts/user-script-plugin-document.html because it is passing now.

* platform/win/Skipped:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71434 268f45cc-cd09-0410-ab3c-d52691b4dbfc