Switch back to using kCGImageAlphaPremultipliedFirst when generating pixel dumps on Windows
I changed this behavior in r71418 thinking that it was required for
getting plugins to show up in pixel dumps. But it doesn't seem to be
necessary, and was making it impossible to compare new Windows pixel
dumps with existing Windows or Mac pixel dumps (because ImageDiff won't
compare an image with alpha to an image without alpha).

Fixes <http://webkit.org/b/49172> REGRESION (r71418): Can't compare
new Windows pixel results to existing Windows or Mac results

Reviewed by Antti Koivisto.

* DumpRenderTree/win/PixelDumpSupportWin.cpp:
(createBitmapContextFromWebView): Replaced kCGImageAlphaNoneSkipFirst
with kCGImageAlphaPremultipliedFirst.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@71516 268f45cc-cd09-0410-ab3c-d52691b4dbfc